# docs/reference/openapi.yaml
openapi: 3.1.0
info:
  title: Viaduct Operator API
  version: v2
  description: >
    Stable Viaduct operator and automation endpoints. This document covers the
    release-grade API surfaces used by the packaged dashboard, CLI-assisted
    workflows, and tenant automation. It is additive and may omit internal or
    less frequently used route variants.
servers:
  - url: http://localhost:8080
paths:
...
  /api/v1/snapshots:
    get:
      summary: List tenant discovery snapshots
      security:
        - TenantAPIKey: []
        - ServiceAccountKey: []
      responses:
        "200":
          description: Snapshot metadata ordered from newest to oldest
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SnapshotMeta"
...
  /api/v2/snapshots:
    get:
      summary: List tenant discovery snapshots (paginated)
      security:
        - TenantAPIKey: []
        - ServiceAccountKey: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: per_page
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        "200":
          description: Snapshot metadata ordered from newest to oldest with pagination metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SnapshotListResponse"
...
  /api/v1/migrations:
    get:
      summary: List tenant migrations
      security:
        - TenantAPIKey: []
        - ServiceAccountKey: []
      responses:
        "200":
          description: Migration history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MigrationMeta"
    post:
      summary: Create or plan a migration
      security:
        - TenantAPIKey: []
        - ServiceAccountKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "202":
          description: Migration plan accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MigrationState"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "403":
          $ref: "#/components/responses/ErrorResponse"
        "500":
          $ref: "#/components/responses/ErrorResponse"
  /api/v2/migrations:
    get:
      summary: List tenant migrations (paginated)
      security:
        - TenantAPIKey: []
        - ServiceAccountKey: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: per_page
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        "200":
          description: Migration history with pagination metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MigrationListResponse"
    post:
      summary: Create or plan a migration
      security:
        - TenantAPIKey: []
        - ServiceAccountKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "202":
          description: Migration plan accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MigrationState"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "403":
          $ref: "#/components/responses/ErrorResponse"
        "500":
          $ref: "#/components/responses/ErrorResponse"
  /api/v1/migrations/{migrationID}:
    get:
      summary: Get migration state
      security:
        - TenantAPIKey: []
        - ServiceAccountKey: []
      parameters:
        - in: path
          name: migrationID
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Migration state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MigrationState"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "403":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
        "500":
          $ref: "#/components/responses/ErrorResponse"
...
  /api/v1/reports/{name}:
    get:
      summary: Export tenant report
      security:
        - TenantAPIKey: []
        - ServiceAccountKey: []
      parameters:
        - in: path
          name: name
          required: true
          schema:
            type: string
            enum: [summary, migrations, audit]
        - in: query
          name: format
          schema:
            type: string
            enum: [json, csv]
      responses:
        "200":
          description: Requested report
          content:
            application/json:
              schema:
                type: object
            text/csv:
              schema:
                type: string
  /api/v1/workspaces:
    get:
      summary: List pilot workspaces
      security:
        - TenantAPIKey: []
        - ServiceAccountKey: []
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: Saved pilot workspaces
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PilotWorkspace"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "403":
          $ref: "#/components/responses/ErrorResponse"
        "500":
          $ref: "#/components/responses/ErrorResponse"
    post:
      summary: Create a pilot workspace
      security:
        - TenantAPIKey: []
        - ServiceAccountKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkspaceCreateRequest"
      responses:
        "201":
          description: Pilot workspace created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PilotWorkspace"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "403":
          $ref: "#/components/responses/ErrorResponse"
        "409":
          $ref: "#/components/responses/ErrorResponse"
        "500":
          $ref: "#/components/responses/ErrorResponse"
  /api/v1/workspaces/{workspaceID}:
    get:
      summary: Get a pilot workspace
      security:
        - TenantAPIKey: []
        - ServiceAccountKey: []
      parameters:
        - in: path
          name: workspaceID
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Pilot workspace state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PilotWorkspace"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "403":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
        "500":
          $ref: "#/components/responses/ErrorResponse"
    delete:
      summary: Delete a pilot workspace
      security:
        - TenantAPIKey: []
        - ServiceAccountKey: []
      parameters:
        - in: path
          name: workspaceID
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Pilot workspace deleted
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "403":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
        "500":
          $ref: "#/components/responses/ErrorResponse"
    patch:
      summary: Update pilot workspace state
      security:
        - TenantAPIKey: []
        - ServiceAccountKey: []
      parameters:
        - in: path
          name: workspaceID
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkspaceUpdateRequest"
      responses:
        "200":
          description: Updated pilot workspace state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PilotWorkspace"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "403":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
        "500":
          $ref: "#/components/responses/ErrorResponse"
...
components:
  securitySchemes:
    TenantAPIKey:
      type: apiKey
      in: header
      name: X-API-Key
    ServiceAccountKey:
      type: apiKey
      in: header
      name: X-Service-Account-Key
    AdminAPIKey:
      type: apiKey
      in: header
      name: X-Admin-Key
  responses:
    ErrorResponse:
      description: Structured API error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorEnvelope"
  schemas:
...
    ApiErrorEnvelope:
      type: object
      properties:
        error:
          $ref: "#/components/schemas/ApiError"
...
    Pagination:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        per_page:
          type: integer
        total_pages:
          type: integer
...
    SnapshotMeta:
      type: object
      properties:
        id:
          type: string
        source:
          type: string
        platform:
          type: string
        vm_count:
          type: integer
        discovered_at:
          type: string
          format: date-time
    SnapshotListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/SnapshotMeta"
        pagination:
          $ref: "#/components/schemas/Pagination"
...
    MigrationMeta:
      type: object
      properties:
        id:
          type: string
        spec_name:
          type: string
        phase:
          type: string
        started_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
    MigrationListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/MigrationMeta"
        pagination:
          $ref: "#/components/schemas/Pagination"
...
    MigrationCheckpoint:
      type: object
      properties:
        phase:
          type: string
        status:
          type: string
        message:
          type: string
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
        diagnostics:
          type: array
          items:
            type: string
    ApprovalGate:
      type: object
      properties:
        required:
          type: boolean
        approved_by:
          type: string
        approved_at:
          type: string
          format: date-time
        ticket:
          type: string
    MigrationPlan:
      type: object
      properties:
        requires_approval:
          type: boolean
        approval_satisfied:
          type: boolean
        waves:
          type: array
          items:
            type: object
            additionalProperties: true
...
    MigrationState:
      type: object
      properties:
        id:
          type: string
        phase:
          type: string
        started_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
        errors:
          type: array
          items:
            type: string
        checkpoints:
          type: array
          items:
            $ref: "#/components/schemas/MigrationCheckpoint"
        approval:
          $ref: "#/components/schemas/ApprovalGate"
        pending_approval:
          type: boolean
        plan:
          $ref: "#/components/schemas/MigrationPlan"
...
    Platform:
      type: string
      enum: [vmware, proxmox, hyperv, kvm, nutanix]
    PilotWorkspaceStatus:
      type: string
      enum: [draft, discovered, graph-ready, simulated, planned, reported]
...
    WorkspaceSourceConnection:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        platform:
          $ref: "#/components/schemas/Platform"
        address:
          type: string
        credential_ref:
          type: string
        last_snapshot_id:
          type: string
        last_discovered_at:
          type: string
          format: date-time
    WorkspaceSnapshot:
      type: object
      properties:
        snapshot_id:
          type: string
        source_connection_id:
          type: string
        source:
          type: string
        platform:
          $ref: "#/components/schemas/Platform"
        vm_count:
          type: integer
        discovered_at:
          type: string
          format: date-time
    WorkspaceTargetAssumptions:
      type: object
      properties:
        platform:
          $ref: "#/components/schemas/Platform"
        address:
          type: string
        credential_ref:
          type: string
        default_host:
          type: string
        default_storage:
          type: string
        default_network:
          type: string
        notes:
          type: string
    WorkspacePlanSettings:
      type: object
      properties:
        name:
          type: string
        parallel:
          type: integer
        verify_boot:
          type: boolean
        approval_required:
          type: boolean
        approved_by:
          type: string
        approval_ticket:
          type: string
        window_start:
          type: string
          format: date-time
        window_end:
          type: string
          format: date-time
        wave_size:
          type: integer
        dependency_aware:
          type: boolean
    WorkspaceGraphArtifact:
      type: object
      properties:
        generated_at:
          type: string
          format: date-time
        node_count:
          type: integer
        edge_count:
          type: integer
        raw_json:
          type: object
          additionalProperties: true
    WorkspaceSimulationArtifact:
      type: object
      properties:
        generated_at:
          type: string
          format: date-time
        target_platform:
          $ref: "#/components/schemas/Platform"
        selected_workload_ids:
          type: array
          items:
            type: string
        moved_vms:
          type: integer
        raw_json:
          type: object
          additionalProperties: true
...
    WorkspaceReadinessResult:
      type: object
      properties:
        generated_at:
          type: string
          format: date-time
        status:
          $ref: "#/components/schemas/WorkspaceReadinessStatus"
        selected_workload_count:
          type: integer
        recommendation_count:
          type: integer
        policy_violation_count:
          type: integer
        blocking_issues:
          type: array
          items:
            type: string
        warning_issues:
          type: array
          items:
            type: string
...
    WorkspaceApproval:
      type: object
      properties:
        id:
          type: string
        stage:
          type: string
        status:
          $ref: "#/components/schemas/WorkspaceApprovalStatus"
        approved_by:
          type: string
        ticket:
          type: string
        notes:
          type: string
        created_at:
          type: string
          format: date-time
...
    PilotWorkspace:
      type: object
      properties:
        id:
          type: string
        tenant_id:
          type: string
        name:
          type: string
        description:
          type: string
        status:
          $ref: "#/components/schemas/PilotWorkspaceStatus"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        source_connections:
          type: array
          items:
            $ref: "#/components/schemas/WorkspaceSourceConnection"
        snapshots:
          type: array
          items:
            $ref: "#/components/schemas/WorkspaceSnapshot"
        selected_workload_ids:
          type: array
          items:
            type: string
        target_assumptions:
          $ref: "#/components/schemas/WorkspaceTargetAssumptions"
        plan_settings:
          $ref: "#/components/schemas/WorkspacePlanSettings"
        graph:
          $ref: "#/components/schemas/WorkspaceGraphArtifact"
        simulation:
          $ref: "#/components/schemas/WorkspaceSimulationArtifact"
        readiness:
          $ref: "#/components/schemas/WorkspaceReadinessResult"
        saved_plan:
          $ref: "#/components/schemas/WorkspaceSavedPlan"
        approvals:
          type: array
          items:
            $ref: "#/components/schemas/WorkspaceApproval"
        notes:
          type: array
          items:
            $ref: "#/components/schemas/WorkspaceNote"
        reports:
          type: array
          items:
            $ref: "#/components/schemas/WorkspaceReportArtifact"
    WorkspaceCreateRequest:
      type: object
      required: [name]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        source_connections:
          type: array
          items:
            $ref: "#/components/schemas/WorkspaceSourceConnection"
        selected_workload_ids:
          type: array
          items:
            type: string
        target_assumptions:
          $ref: "#/components/schemas/WorkspaceTargetAssumptions"
        plan_settings:
          $ref: "#/components/schemas/WorkspacePlanSettings"
        approvals:
          type: array
          items:
            $ref: "#/components/schemas/WorkspaceApproval"
        notes:
          type: array
          items:
            $ref: "#/components/schemas/WorkspaceNote"
    WorkspaceUpdateRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        source_connections:
          type: array
          items:
            $ref: "#/components/schemas/WorkspaceSourceConnection"
        selected_workload_ids:
          type: array
          items:
            type: string
        target_assumptions:
          $ref: "#/components/schemas/WorkspaceTargetAssumptions"
        plan_settings:
          $ref: "#/components/schemas/WorkspacePlanSettings"
        approvals:
          type: array
          items:
            $ref: "#/components/schemas/WorkspaceApproval"
        notes:
          type: array
          items:
            $ref: "#/components/schemas/WorkspaceNote"
...